requirements.txt
numpy
tensorflow
paho-mqtt
